requirements.txt
aws-cdk-lib>=2.178.1
constructs>=10.4.2
python-dotenv>=1.0.1